+++ dotNet/GoT/GoT.Server/Messages/HouseCardDto.cs
@@ -7,12 +7,12 @@ namespace GoT.Server.Messages
 {
     public class HouseCardDto
     {
-        public Int64 HouseCardId { get; set; }
+        public Int32 HouseCardId { get; set; }
         public string Description { get; set; }
         public Int32 CombatStrength { get; set; }
         public Int32 SwordCount { get; set; }
         public Int32 TowerCount { get; set; }
         public bool IsAvailable { get; set; }
-        public Int64 CharacterId { get; set; }
+        public Int32 CharacterId { get; set; }
     }
 }
